Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,6 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -78,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
@@ -86,11 +94,16 @@ class QuestionsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Question  $question
+     * @param  \App\Question $question
      * @return \Illuminate\Http\Response
+     * @throws \Exception
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
